// apps/api/src/orders/order-status.enum.ts
export enum MasterOrderStatus {
  PENDING_PAYMENT = "PENDING_PAYMENT",
  CONFIRMED = "CONFIRMED",
  PREPARING = "PREPARING",
  PARTIALLY_READY = "PARTIALLY_READY",
  PARTIALLY_FULFILLED = "PARTIALLY_FULFILLED",
  READY = "READY",
  OUT_FOR_DELIVERY = "OUT_FOR_DELIVERY",
  DELIVERED = "DELIVERED",
  CANCELLED = "CANCELLED",
}

export enum SubOrderStatus {
  PENDING = "PENDING",
  ACCEPTED = "ACCEPTED",
  PREPARING = "PREPARING",
  READY = "READY",
  COLLECTED = "COLLECTED",
  DISPATCHED = "DISPATCHED",
  REJECTED = "REJECTED",
}